Refactor and split switch to methods into new, separate iframe group and add to window group.
