+++ src/browserist/browser/window/__main__.py
@@ -6,11 +6,13 @@ from .get.__main__ import WindowGetDriverMethods
 from .maximize import window_maximize
 from .minimize import window_minimize
 from .set.__main__ import WindowSetDriverMethods
+from .switch_to import switch_to_window
 
 
 class WindowDriverMethods(DriverMethods):
-    def __init__(self, browser_driver: BrowserDriver, settings: BrowserSettings) -> None:
+    def __init__(self, browser_driver: BrowserDriver, settings: BrowserSettings, original_window_handle: str) -> None:
         super().__init__(browser_driver, settings)
+        self._original_window_handle = original_window_handle
         self.get: WindowGetDriverMethods = WindowGetDriverMethods(browser_driver, settings)
         self.set: WindowSetDriverMethods = WindowSetDriverMethods(browser_driver, settings)
 
@@ -28,3 +30,13 @@ class WindowDriverMethods(DriverMethods):
         """Minimizes the window of current browsing context. The exact behavior of this command is specific to individual window managers. Minimize Window typically hides the window in the system tray."""
 
         window_minimize(self._driver)
+
+    def switch_to(self, window_handle: str) -> None:
+        """Switch to window by handle ID."""
+
+        switch_to_window(self._driver, window_handle)
+
+    def switch_to_original_window(self) -> None:
+        """Switch to initial window."""
+
+        switch_to_window(self._driver, self._original_window_handle)
