+++ src/browserist/browser/window/switch_to.py
